Recursion error appeared to be due to the use of the summation() function in XXXvibrational_matrix_element()XXX Term.sympy().  Replaced it with the pythonic approach and currently running test for <Nu22| Ht22 | Nu33>.
Test case running for > 48 hours, presumably the evaluation of the sympy expressions generated in the intermediate sums was causing the significant time length.  Replacing with evaluation of individual permutations of the indices instead of expanding one index at a time into a sympy expression.

Not sure if this is faster than the previous approach, but regardless it is still too slow.  Timing tests for <Nu22 | Ht22 | Nu33> estimates 140 hours to complete.  I believe the sympy overhead is just too much for the ~3 000 000 iterations to evaluate.  Will work on creating a way to export the equation definitions to a python script for more direct evaluation via functions.